+++ bank/schemas.py
@@ -4,7 +4,7 @@ from marshmallow import Schema, fields, validate
 class ApplicationSchema(Schema):
     id = fields.Integer(dump_only=True)
     user_id = fields.Integer()
-    value = fields.Float()
+    value = fields.Integer()
     request_date = fields.DateTime()
     answer_date = fields.DateTime()
     approved = fields.Boolean()
@@ -16,6 +16,6 @@ class UserSchema(Schema):
     first_name = fields.String(validate=[validate.Length(max=250)])
     last_name = fields.String(validate=[validate.Length(max=250)])
     username = fields.String(validate=[validate.Length(max=250)])
-    debt = fields.Float()
+    debt = fields.Integer()
     # applications = fields.Nested(ApplicationSchema, many=True, dump_only=True)
     # message = fields.String(dump_only=True)
